+++ cmd/apps/skysocks-client/skysocks-client.go
@@ -14,7 +14,7 @@ import (
 
 	"github.com/SkycoinProject/skywire-mainnet/internal/netutil"
 	"github.com/SkycoinProject/skywire-mainnet/internal/skyenv"
-	"github.com/SkycoinProject/skywire-mainnet/internal/therealproxy"
+	"github.com/SkycoinProject/skywire-mainnet/internal/skysocks"
 	"github.com/SkycoinProject/skywire-mainnet/pkg/app"
 	"github.com/SkycoinProject/skywire-mainnet/pkg/app/appnet"
 	"github.com/SkycoinProject/skywire-mainnet/pkg/routing"
@@ -30,7 +30,7 @@ var r = netutil.NewRetrier(time.Second, 0, 1)
 
 func main() {
 	log := app.NewLogger(appName)
-	therealproxy.Log = log.PackageLogger("therealproxy")
+	skysocks.Log = log.PackageLogger("skysocks")
 
 	var addr = flag.String("addr", skyenv.SkyproxyClientAddr, "Client address to listen on")
 	var serverPK = flag.String("srv", "", "PubKey of the server to connect to")
@@ -74,7 +74,7 @@ func main() {
 
 	log.Printf("Connected to %v\n", pk)
 
-	client, err := therealproxy.NewClient(conn)
+	client, err := skysocks.NewClient(conn)
 	if err != nil {
 		log.Fatal("Failed to create a new client: ", err)
 	}
