+++ go.mod
@@ -12,6 +12,8 @@ require (
 	github.com/google/uuid v1.1.1
 	github.com/gorilla/handlers v1.4.2
 	github.com/gorilla/securecookie v1.1.1
+	github.com/hashicorp/yamux v0.0.0-20190923154419-df201c70410d
+	github.com/mitchellh/go-homedir v1.1.0
 	github.com/pkg/profile v1.3.0
 	github.com/prometheus/client_golang v1.2.1
 	github.com/prometheus/common v0.7.0
