+++ internal/skysocks/server_test.go
@@ -1,4 +1,4 @@
-package therealproxy
+package skysocks
 
 import (
 	"fmt"
@@ -11,6 +11,7 @@ import (
 	"time"
 
 	"github.com/SkycoinProject/skycoin/src/util/logging"
+	"github.com/hashicorp/yamux"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"golang.org/x/net/nettest"
@@ -49,8 +50,12 @@ func TestProxy(t *testing.T) {
 	conn, err := net.Dial("tcp", l.Addr().String())
 	require.NoError(t, err)
 
-	client, err := NewClient(conn)
+	session, err := yamux.Client(conn, nil)
 	require.NoError(t, err)
+
+	client := &Client{
+		session: session,
+	}
 
 	errChan2 := make(chan error)
 	go func() {
